create register middleware update user model (tg_id(Integer -> bigint) delete method welcome user in start command

+++ bot/database/user.py
@@ -1,6 +1,6 @@
 import datetime
 
-from sqlalchemy import Column, VARCHAR, INTEGER, DATE
+from sqlalchemy import Column, VARCHAR, INTEGER, DATE, BigInteger
 from sqlalchemy.orm import relationship
 
 from .base import BaseModel
@@ -11,5 +11,5 @@ class User(BaseModel):
     user_id = Column(INTEGER, primary_key=True, unique=True, nullable=False, autoincrement=True)
     user_name = Column(VARCHAR(20), unique=False, nullable=True)
     reg_date = Column(DATE, default=datetime.date.today())
-    tg_id = Column(INTEGER, unique=True)
+    tg_id = Column(BigInteger, unique=True)
     employees = relationship("Employee", back_populates="user")
